🧹 Extract UI components and types from PaperDetailClient
