+++ apps/web/src/app/papers/[id]/__tests__/page-metadata.test.tsx
@@ -91,11 +91,9 @@ describe("papers/[id]/page metadata", () => {
     expect(metadata.openGraph?.title).toBe("成果物詳細 | OpenShelf");
   });
 
-
   it("generates generic metadata when id is invalid", async () => {
     // Generate metadata without await as the component is sync for generating metadata via props
     const metadata = await generateMetadata({ params: { id: "../bad" } });
     expect(metadata.title).toBe("成果物詳細 | OpenShelf");
   });
-
 });
