+++ apps/web/src/app/papers/[id]/__tests__/paper-detail-client.test.tsx
@@ -417,7 +417,9 @@ describe("PaperDetailClient", () => {
     const slideRow = screen.getByText("deck.pptx").closest("li");
     expect(slideRow).not.toBeNull();
     fireEvent.click(
-      within(slideRow!).getByRole("button", { name: "deck.pptxをダウンロード" }),
+      within(slideRow!).getByRole("button", {
+        name: "deck.pptxをダウンロード",
+      }),
     );
 
     await waitFor(() => {
@@ -748,18 +750,21 @@ describe("PaperDetailClient", () => {
     [401, "ログインが必要です"],
     [404, "成果物が見つかりません"],
     [500, "成果物の取得に失敗しました"],
-  ])("maps paper fetch status %s to its error message", async (status, message) => {
-    vi.mocked(apiFetch).mockResolvedValue(new Response("error", { status }));
+  ])(
+    "maps paper fetch status %s to its error message",
+    async (status, message) => {
+      vi.mocked(apiFetch).mockResolvedValue(new Response("error", { status }));
 
-    render(
-      <PaperDetailClient
-        paperId="paper-1"
-        siteBase="https://openshelf.example"
-      />,
-    );
+      render(
+        <PaperDetailClient
+          paperId="paper-1"
+          siteBase="https://openshelf.example"
+        />,
+      );
 
-    expect(await screen.findByText(message)).toBeInTheDocument();
-  });
+      expect(await screen.findByText(message)).toBeInTheDocument();
+    },
+  );
 
   it("shows an error when API request fails entirely", async () => {
     vi.mocked(apiFetch).mockRejectedValue(new Error("Network Error"));
@@ -775,5 +780,4 @@ describe("PaperDetailClient", () => {
       await screen.findByText("成果物の取得に失敗しました"),
     ).toBeInTheDocument();
   });
-
 });
