+++ apps/web/src/app/papers/[id]/cite-button.tsx
@@ -121,9 +121,7 @@ export function CiteButton({ paperId }: CiteButtonProps) {
                 <span className="flex items-center gap-2">
                   <Spinner className="h-3 w-3" />
                   <span aria-hidden="true">生成中...</span>
-                  <span className="sr-only">
-                    {option.label} の引用を生成中
-                  </span>
+                  <span className="sr-only">{option.label} の引用を生成中</span>
                 </span>
               ) : (
                 option.label
